changed int to long data type

+++ src/main/java/woodmen/Application/Customer.java
@@ -2,21 +2,21 @@ package Woodmen.Application;
 
 public class Customer {
 
-    private int id;
+    private Long id;
     private String firstName;
     private String lastName;
 
-    public Customer(String firstName, String lastName, int id){
+    public Customer(String firstName, String lastName, Long id){
         this.firstName = firstName;
         this.lastName = lastName;
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
